requirements.txt
streamlit==1.32.0
yfinance==0.2.37
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
pillow==10.2.0